+++ noodles-vcf/src/record/builder.rs
@@ -243,7 +243,7 @@ impl Builder {
     /// ```
     /// use noodles_vcf::{
     ///     self as vcf,
-    ///     record::{info::field::{key, value::Array, Value}, Info, Position},
+    ///     record::{info::field::{key, Value}, Info, Position},
     /// };
     ///
     /// let record = vcf::Record::builder()
@@ -256,7 +256,7 @@ impl Builder {
     ///
     /// let expected = [
     ///     (key::SAMPLES_WITH_DATA_COUNT, Some(Value::Integer(3))),
-    ///     (key::ALLELE_FREQUENCIES, Some(Value::Array(Array::Float(vec![Some(0.5)])))),
+    ///     (key::ALLELE_FREQUENCIES, Some(Value::from(vec![Some(0.5)]))),
     /// ].into_iter().collect();
     ///
     /// assert_eq!(record.info(), &expected);